+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add support for tilde(~) and '${homedir}' in alternative configuration filenames
+
 # v0.0.20 - 2019-10-16
 - Warn if configuration file is not found when formatting
 - Tweak setting of range ends